requirements.txt
PyYAML~=6.0.1
yacs~=0.1.8
pandas~=2.1.3
requests~=2.31.0